Add some tests

+++ BooksApi/Models/Book.cs
@@ -2,10 +2,10 @@
 {
     public class Book
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Title { get; set; }
 
-        public  string Author { get; set; }
+        public string Author { get; set; }
     }
 }
